+++ CHANGELOG.md
@@ -4,7 +4,9 @@ This list is not intended to be all-encompassing - it will document major and br
 changes with their rationale when appropriate:
 
 ### v5.6.4.0 (uncut)
+- **http4k-*** : Upgrade some dependency versions
 - **http4k-server-apache4**: upgrade compromised `commons-codec` version  H/T @oharaandrew314
+- **http4k-template-jte** : [New module] JTE templating support
 
 ### v5.6.3.0
 - **http4k-security-oauth** :Add ability to override response mode in OAuthProvider.